splash-service/requirements.txt
torch==2.0.1
torchvision==0.15.2
diffusers==0.19.3
transformers==4.31.0
accelerate==0.21.0
fastapi==0.68.1
uvicorn[standard]==0.15.0
python-multipart==0.0.5
pillow==8.3.2
numpy==1.24.3
opencv-python-headless==4.8.0.74
redis==4.0.2
prometheus-client==0.11.0
opentelemetry-api==1.7.1
opentelemetry-sdk==1.7.1
pydantic==1.8.2
python-dotenv==0.19.0 